// src/payload-types.ts
export interface User {
  id: string;
  name?: string | null;
  birthday?: string | null;
  bio?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  avatar?: (string | null) | Media;
  roles?: ('admin' | 'user' | 'super-admin')[] | null;
  families?: {
    docs?: (string | Family)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  lists?: {
    docs?: (string | List)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  invites?: {
    docs?: (string | Invite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "family".
 */
export interface Family {
  id: string;
  name: string;
  description?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  members?: (string | User)[] | null;
  allowInvites?: boolean | null;
  managers?: (string | User)[] | null;
  createdBy?: (string | null) | User;
  invites?: {
    docs?: (string | Invite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  inviteLinkToken?: string | null;
  inviteLinkExpiry?: string | null;
  requireApproval?: boolean | null;
  events?: {
    docs?: (string | Event)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  favorite?: {
    docs?: (string | Favorite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "invite".
 */
export interface Invite {
  id: string;
  email?: string | null;
  token: string;
  tokenExpiry: string;
  eventResponse?: ('declined' | 'maybe' | 'accepted') | null;
  needsApproval?: boolean | null;
  approvalRejected?: boolean | null;
  family?: (string | null) | Family;
  event?: (string | null) | Event;
  user?: (string | null) | User;
  inviter?: (string | null) | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "event".
 */
export interface Event {
  id: string;
  name: string;
  date?: string | null;
  time?: string | null;
  endDate?: string | null;
  endTime?: string | null;
  location?: string | null;
  info?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  family?: (string | null) | Family;
  creator: string | User;
  managers?: (string | User)[] | null;
  invites?: {
    docs?: (string | Invite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  secretSantaBudget?: string | null;
  secretSantaParticipants?: (string | Invite)[] | null;
  assignments?: {
    docs?: (string | Assignment)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  exclusions?: {
    docs?: (string | Exclusion)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  secretSantaNotificationsSent?: boolean | null;
  favorite?: {
    docs?: (string | Favorite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Favorite {
  id: string;
  user: string | User;
  favorite?:
    | ({
        relationTo: 'family';
        value: string | Family;
      } | null)
    | ({
        relationTo: 'event';
        value: string | Event;
      } | null)
    | ({
        relationTo: 'list';
        value: string | List;
      } | null);
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "list".
 */
export interface List {
  id: string;
  name: string;
  description?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  user: string | User;
  items?: (string | Item)[] | null;
  public?: boolean | null;
  visibleToFamilies?: (string | Family)[] | null;
  visibleToUsers?: (string | User)[] | null;
  shareLink?: string | null;
  favorite?: {
    docs?: (string | Favorite)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "item".
 */
export interface Item {
  id: string;
  name: string;
  list: {
    docs?: (string | List)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  link?: string | null;
  notes?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  priority?: ('low' | 'medium' | 'high') | null;
  price?: string | null;
  image?: (string | null) | Media;
  purchasedBy?: (string | User)[] | null;
  updatedAt: string;
  createdAt: string;
}